requirements.txt
torch>=1.9.0+cpu
torchvision>=0.10.0+cpu
matplotlib>=3.3.0
numpy>=1.19.0
Pillow>=8.0.0
fastapi>=0.100.0
uvicorn[standard]>=0.22.0
python-multipart>=0.0.6
requests>=2.25.0